Refactor loot table class location
